+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/setup/PostSetup.java
@@ -5,6 +5,8 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -34,6 +36,7 @@ import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Lists.SlimefunItems;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AContainer;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.MachineRecipe;
 import me.mrCookieSlime.Slimefun.api.Slimefun;
 
 public final class PostSetup {
@@ -190,7 +193,8 @@ public final class PostSetup {
 
     private static void loadSmelteryRecipes() {
         Smeltery smeltery = (Smeltery) SlimefunItems.SMELTERY.getItem();
-        if (smeltery != null) {
+
+        if (smeltery != null && !smeltery.isDisabled()) {
             ItemStack[] input = null;
 
             for (ItemStack[] recipe : smeltery.getRecipes()) {
@@ -220,6 +224,13 @@ public final class PostSetup {
                     }
 
                     input = null;
+                }
+            }
+
+            for (SlimefunItem item : SlimefunPlugin.getRegistry().getEnabledSlimefunItems()) {
+                if (item instanceof AContainer && ((AContainer) item).getMachineIdentifier().equals("ELECTRIC_SMELTERY")) {
+                    List<MachineRecipe> recipes = ((AContainer) item).getMachineRecipes();
+                    Collections.sort(recipes, Comparator.comparingInt(recipe -> recipe == null ? 0 : -recipe.getInput().length));
                 }
             }
         }