+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/abstractItems/AContainer.java
@@ -165,6 +165,10 @@ public abstract class AContainer extends SlimefunItem implements InventoryBlock,
         // Override this method to register your machine recipes
     }
 
+    public List<MachineRecipe> getMachineRecipes() {
+        return recipes;
+    }
+
     public List<ItemStack> getDisplayRecipes() {
         List<ItemStack> displayRecipes = new ArrayList<>(recipes.size() * 2);